Remove unnecessary comments, add pre-commit config, delete unused functions in cmd/podman, delete unused config folder (maybe will add later)
