+++ src/podman/configure.go
@@ -2,27 +2,24 @@ package podman
 
 import (
 	"fmt"
-	"infra-lab-cli/config"
 	"infra-lab-cli/utils"
 	"os/exec"
 	"strconv"
 )
 
-// ConfigureParams holds the configuration parameters for a machine
-type ConfigureParams struct {
-	CPUs     string
-	Memory   string
-	DiskSize string
+func isParamChanged(param string, currentValue string) bool {
+	if param != "0" && param != "" {
+		return param != currentValue
+	}
+	return false
 }
 
-// ConfigureMachine configures the specified podman machine with the given parameters
 func ConfigureMachine(binaryName, machineName string, params ConfigureParams) error {
-	if !config.IsBinaryInPath(binaryName) {
-		fmt.Print(config.BinaryNotFoundError(binaryName))
+	if !utils.IsBinaryInPath(binaryName) {
+		fmt.Print(utils.BinaryNotFoundError(binaryName))
 		return nil
 	}
 
-	// Get current machine state
 	machines, err := InspectMachine(machineName)
 	if err != nil {
 		return err
@@ -34,14 +31,12 @@ func ConfigureMachine(binaryName, machineName string, params ConfigureParams) er
 
 	machine := machines[0]
 
-	// Check if any configuration has changed
 	isChanged := isConfigChanged(params, machine)
 	if !isChanged {
 		fmt.Println("No changes detected in configuration.")
 		return nil
 	}
 
-	// Stop machine if running
 	wasRunning := machine.State == "running"
 	if wasRunning {
 		err = StopMachine(binaryName, machineName)
@@ -50,7 +45,6 @@ func ConfigureMachine(binaryName, machineName string, params ConfigureParams) er
 		}
 	}
 
-	// Update CPU configuration
 	if isParamChanged(params.CPUs, strconv.Itoa(machine.Resources.CPUs)) {
 		out, err := exec.Command("podman", "machine", "set", "--cpus", params.CPUs).CombinedOutput()
 		fmt.Print(string(out))
@@ -60,7 +54,6 @@ func ConfigureMachine(binaryName, machineName string, params ConfigureParams) er
 		fmt.Printf("CPU was updated from %d to %s\n", machine.Resources.CPUs, params.CPUs)
 	}
 
-	// Update memory configuration
 	memMiB, err := utils.ConvertToMiB(params.Memory)
 	if err == nil && isParamChanged(memMiB, strconv.Itoa(machine.Resources.Memory)) {
 		out, err := exec.Command("podman", "machine", "set", "--memory", memMiB).CombinedOutput()
@@ -71,7 +64,6 @@ func ConfigureMachine(binaryName, machineName string, params ConfigureParams) er
 		fmt.Printf("Memory was updated from %d to %s\n", machine.Resources.Memory, params.Memory)
 	}
 
-	// Update disk size configuration
 	if isParamChanged(params.DiskSize, strconv.Itoa(machine.Resources.DiskSize)) {
 		// TODO: new size must be greater than current
 		out, err := exec.Command("podman", "machine", "set", "--disk-size", params.DiskSize).CombinedOutput()
@@ -82,7 +74,6 @@ func ConfigureMachine(binaryName, machineName string, params ConfigureParams) er
 		fmt.Printf("Disk size was updated from %d to %s\n", machine.Resources.DiskSize, params.DiskSize)
 	}
 
-	// Restart machine if it was running
 	if wasRunning {
 		err = StartMachine(binaryName, machineName)
 		if err != nil {
@@ -93,20 +84,6 @@ func ConfigureMachine(binaryName, machineName string, params ConfigureParams) er
 	return nil
 }
 
-// IsParamChanged checks if a parameter has changed from its current value
-func IsParamChanged(param string, currentValue string) bool {
-	if param != "0" && param != "" {
-		return param != currentValue
-	}
-	return false
-}
-
-// isParamChanged is an internal helper function
-func isParamChanged(param string, currentValue string) bool {
-	return IsParamChanged(param, currentValue)
-}
-
-// isConfigChanged checks if any configuration parameter has changed
 func isConfigChanged(params ConfigureParams, machine InspectedMachine) bool {
 	if isParamChanged(params.CPUs, strconv.Itoa(machine.Resources.CPUs)) {
 		return true
