+++ src/podman/start.go
@@ -2,14 +2,13 @@ package podman
 
 import (
 	"fmt"
-	"infra-lab-cli/config"
+	"infra-lab-cli/utils"
 	"os/exec"
 )
 
-// StartMachine starts the specified podman machine.
 func StartMachine(binaryName, machineName string) error {
-	if !config.IsBinaryInPath(binaryName) {
-		fmt.Print(config.BinaryNotFoundError(binaryName))
+	if !utils.IsBinaryInPath(binaryName) {
+		fmt.Print(utils.BinaryNotFoundError(binaryName))
 		return nil
 	}
