+++ src/podman/status.go
@@ -2,14 +2,12 @@ package podman
 
 import (
 	"fmt"
-	"infra-lab-cli/config"
 	"infra-lab-cli/utils"
 )
 
-// GetMachineStatus retrieves and displays the status of the specified podman machine
 func GetMachineStatus(binaryName, machineName string) error {
-	if !config.IsBinaryInPath(binaryName) {
-		fmt.Print(config.BinaryNotFoundError(binaryName))
+	if !utils.IsBinaryInPath(binaryName) {
+		fmt.Print(utils.BinaryNotFoundError(binaryName))
 		return nil
 	}
